Refactor ui code into ui.ts
